pyproject.toml
[project]
name = "fpl-optimizer"
version = "0.1.0"
description = "Fantasy Premier League Weekly Points Optimizer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.0.3,<2.3",
  "numpy>=1.24",
  "requests>=2.32",
  "pyyaml>=6.0",
  "rich>=13.7",
  "ortools>=9.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "black>=24.3",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E","F","I","B","UP","C90"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
